reorganized more code
