+++ contracts/distribution/dao-rewards-distributor/src/lib.rs
@@ -2,6 +2,7 @@
 
 pub mod contract;
 mod error;
+pub mod helpers;
 pub mod hooks;
 pub mod msg;
 pub mod rewards;
